Revert "Change the Dijkstra algorithm to use the networkx library"
This reverts commit a032011d56d11af3d4261895cbcb49d5d26d1f33.
